Add fix for reference URL calculation and add/change coresponding tests
Change the used JSON node for TLP Label from `/csaf/document/distribution/tlp` to `/csaf/document/distribution/tlp/label` and lowercase the label in `calculateReferenceUrl`. This assures that the generated URL contains the correct label and is consistant with the document tree setup by CSAF Publisher.

The `setFinalTrackingIdTest` is changed accoding to the lowercase change and a new `setFinalTrackingIdTest_label_not_null` test is added to test the correct usage of a given label.